==> old_notebooks_1/16-plot_layer_column_plots.ipynb <==
# %load C60_jupyter_imports.py
%load_ext autoreload
%autoreload 2

import xarray as xr
import numpy as np
from dask.distributed import Client
import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.patches as patches

import seaborn as sns
import pandas as pd
from scipy import stats
import xesmf as xe
from scipy.stats import linregress
import os
import requests
import os
import scipy.signal as sps
import cartopy.crs as ccrs
import cartopy.feature as cfeature
import matplotlib.pyplot as plt
# Lets import our functions. There are no tests here. Stored separately so as to not make a mess.
# Make sure in right dir first
#!pip install --user seawater==3.3.4
import seawater

os.chdir('/g/data/xv83/np1383/src_CAFE60_eqpac_analysis_code/')
from C60_obs_functions import convert_trim_fratios
from xarray_trends import xarray_detrend,xarray_get_trend, deseasonaliser
from dask.distributed import Client,Scheduler
from dask_jobqueue import SLURMCluster

use_dask=False

if use_dask==True:
    # Set up the remote dask cluster. Can either use this version or a similar version above if building a LocalCluster.
    
    #cluster = SLURMCluster(cores=2,processes=1,memory="16GB")
    cluster = SLURMCluster(cores=8,processes=1,memory="47GB")#47GB")
    #cluster = SLURMCluster(cores=8,processes=4,memory="22GB")#47GB")
    
    client = Client(cluster)
    cluster.scale(cores=64)
    #cluster.adapt(minimum=2, maximum=16)
    cluster
    #client

# NEW VERSION

region=''#'_big'
dic_transport_through_box = pd.read_csv(f'dic_tx{region}.csv',index_col=0)

#dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(dic_transport_through_box['Upwelling 100 in']+dic_transport_through_box['Upwelling 100 out'])
#dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(dic_transport_through_box['AnthUpwelling 100 in']+dic_transport_through_box['AnthUpwelling 100 out'])

dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(
    dic_transport_through_box['surface_to_upperthermocline in']+dic_transport_through_box['surface_to_upperthermocline out'])
dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(
    dic_transport_through_box['anth_surface_to_upperthermocline in']+dic_transport_through_box['anth_surface_to_upperthermocline out'])

dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(dic_transport_through_box['CO2 flux in']+dic_transport_through_box['CO2 flux out'])
dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(dic_transport_through_box['anth CO2 flux in']+dic_transport_through_box['anth CO2 flux out'])


#dic_transport_through_box['upperthermocline_sum']=dic_transport_through_box['upperthermocline_sum'].fillna(
#   dic_transport_through_box['upper_to_lowerthermocline in']+dic_transport_through_box['upper_to_lowerthermocline out'])
#dic_transport_through_box['upperthermocline_sum']=dic_transport_through_box['upperthermocline_sum'].fillna(
#    dic_transport_through_box['anth_upper_to_lowerthermocline in']+dic_transport_through_box['anth_upper_to_lowerthermocline out'])


dic_transport_through_box['lowerthermocline_sum']=dic_transport_through_box['lowerthermocline_sum'].fillna(
    dic_transport_through_box['lowerthermocline_to_thermostad in']+dic_transport_through_box['lowerthermocline_to_thermostad out'])

dic_transport_through_box['lowerthermocline_sum']=dic_transport_through_box['lowerthermocline_sum'].fillna(
    dic_transport_through_box['anth_lowerthermocline_to_thermostad in']+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])


# Surface now = Anywhere above 24.5

#Total Sum 	Total Sum 100m 	Positive Sum 	Negative Sum

# Total is 600M + and NEG
dic_transport_through_box['Total Sum']=dic_transport_through_box['Total Sum'].fillna(
    dic_transport_through_box['CO2 flux in']+dic_transport_through_box['CO2 flux out'])
dic_transport_through_box['Total Sum']=dic_transport_through_box['Total Sum'].fillna(
    dic_transport_through_box['anth CO2 flux in']+dic_transport_through_box['anth CO2 flux out'])

dic_transport_through_box['Total Sum']=dic_transport_through_box['Total Sum'].fillna(
    dic_transport_through_box['AnthUpwelling 600 in']+dic_transport_through_box['AnthUpwelling 600 out'])
dic_transport_through_box['Total Sum']=dic_transport_through_box['Total Sum'].fillna(
    dic_transport_through_box['Upwelling 600 in']+dic_transport_through_box['Upwelling 600 out'])
    


dic_transport_through_box['Positive Sum']=dic_transport_through_box['Positive Sum'].fillna(
    dic_transport_through_box['CO2 flux in'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])
dic_transport_through_box['Positive Sum']=dic_transport_through_box['Positive Sum'].fillna(
    dic_transport_through_box['anth CO2 flux in'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])

dic_transport_through_box['Positive Sum']=dic_transport_through_box['Positive Sum'].fillna(
    dic_transport_through_box['AnthUpwelling 600 in'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])
dic_transport_through_box['Positive Sum']=dic_transport_through_box['Positive Sum'].fillna(
    dic_transport_through_box['Upwelling 600 in'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])



dic_transport_through_box['Negative Sum']=dic_transport_through_box['Negative Sum'].fillna(
    dic_transport_through_box['CO2 flux out'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])
dic_transport_through_box['Negative Sum']=dic_transport_through_box['Negative Sum'].fillna(
    dic_transport_through_box['anth CO2 flux out'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])

dic_transport_through_box['Negative Sum']=dic_transport_through_box['Negative Sum'].fillna(
    dic_transport_through_box['Upwelling 600 out'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])
dic_transport_through_box['Negative Sum']=dic_transport_through_box['Negative Sum'].fillna(
    dic_transport_through_box['AnthUpwelling 600 out'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])



    
# Need to do some calculations to fill the Positive Sum above thermocline and above thermostad ['Negative Sum']
# Also need to reverse some of the directions (ie E pacific?)
    
dic_transport_through_box=dic_transport_through_box.drop(['Upwelling 100 in','Upwelling 100 out','AnthUpwelling 100 in','AnthUpwelling 100 out'],axis=1)#,'Upwelling 23 out','Upwelling 24.5 in','Upwelling 24.5 out','Upwelling 26.3 in','Upwelling 26.3 out'],axis=1)

dic_transport_through_box=dic_transport_through_box.drop(['Upwelling 600 in','Upwelling 600 out','AnthUpwelling 600 in','AnthUpwelling 600 out'],axis=1)#,'Upwelling 23 out','Upwelling 24.5 in','Upwelling 24.5 out','Upwelling 26.3 in','Upwelling 26.3 out'],axis=1)
dic_transport_through_box=dic_transport_through_box.drop(['surface_to_upperthermocline in','surface_to_upperthermocline out','anth_surface_to_upperthermocline in','anth_surface_to_upperthermocline out','upper_to_lowerthermocline in','upper_to_lowerthermocline out','anth_upper_to_lowerthermocline in','anth_upper_to_lowerthermocline out'],axis=1)
dic_transport_through_box=dic_transport_through_box.drop(['lowerthermocline_to_thermostad in','lowerthermocline_to_thermostad out','anth_lowerthermocline_to_thermostad in','anth_lowerthermocline_to_thermostad out'],axis=1)
dic_transport_through_box=dic_transport_through_box.drop(['CO2 flux in','CO2 flux out','anth CO2 flux in','anth CO2 flux out'],axis=1)
dic_transport_through_box=dic_transport_through_box.dropna(thresh=1)

dic_transport_through_box.to_csv(f'dic_tx2{region}.csv')

t=pd.read_csv(f'dic_tx2{region}.csv').replace(0,np.nan)
merged=t.groupby('Unnamed: 0').mean().sort_values('Unnamed: 0').round(3)#agg(np.mean())

merged['enso_phase']=merged.index.str.split(' ').str[0]
merged.index=merged.index.str.split(' ').str[1:].str.join(' ')#.groupby('surface_sum')

merged.to_csv(f'dic_tx3{region}.csv')
merged



# Prepare Data for Plotting
dic_transport_through_box=pd.read_csv(f'dic_tx3{region}.csv',index_col=0).replace(0,np.nan)

names=dic_transport_through_box.index

new_tx=dic_transport_through_box[~names.str.lower().str.contains('old')]
new_tx_anth=new_tx[new_tx.index.str.lower().str.contains('anth')]
for anth in new_tx_anth.index:
    if anth in new_tx.index:
        new_tx=new_tx.drop(anth)
        
    
# Plotting
plt.figure(figsize=(16,16))
for i,col in enumerate(new_tx.columns[:-1]):
    ax=plt.subplot(5,2,i+1)
    df=new_tx[[col,'enso_phase']]#.plot.bar()
    print(df.pivot(columns='enso_phase').dropna(thresh=1))#.plot#.bar(ax=ax)#plt.show()
    #df.drop('Upwelling 100').groupby('enso_phase').sum().T['EP'].values
    sum_of_rows=df.groupby('enso_phase').sum()#.drop('Upwelling 100').
    #print(sum_of_rows
    plt.title(col)
    plt.xticks(rotation=45)
    plt.legend([f'CP {sum_of_rows.T["CP"].values[0].round(3)} PgAC',
            f'EP {sum_of_rows.T["EP"].values[0].round(3)} PgAC',
            f'Neutral {sum_of_rows.T["Neutral"].values[0].round(3)}PgC',
            f'Nina {sum_of_rows.T["Nina"].values[0].round(3)}PgC'])# Make sure this doesnt break anything
plt.tight_layout()
# Plotting Anth
plt.figure(figsize=(16,16))
for i, col in enumerate(new_tx_anth.columns[:-1]):
    ax=plt.subplot(4,2,i+1)
    df=new_tx_anth[[col,'enso_phase']]#.plot.bar()
    
    
    # CAN WE SOMEHOW MODIFY BAR COLOURS HERE?
    df.pivot(columns='enso_phase').plot.bar(ax=ax)#plt.show()
    #plt.legend(['CP','EP','Neutral','Nina'])# Make sure this doesnt break anything
    plt.title(col)
    sum_of_rows=df.drop('AnthUpwelling 100').groupby('enso_phase').sum()#
    #print(sum_of_rows)
    plt.xticks(rotation=45)
    plt.legend([f'CP {sum_of_rows.T["CP"].values[0].round(3)} PgAC',
                f'EP {sum_of_rows.T["EP"].values[0].round(3)} PgAC',
                f'Neutral {sum_of_rows.T["Neutral"].values[0].round(3)}PgC',
                f'Nina {sum_of_rows.T["Nina"].values[0].round(3)}PgC'])# Make sure this doesnt break anything
plt.tight_layout()


# OLD VERSION

region=''#'_big'
dic_transport_through_box = pd.read_csv(f'dic_tx{region}.csv',index_col=0)

dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(dic_transport_through_box['Upwelling 100 in']+dic_transport_through_box['Upwelling 100 out'])
dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(dic_transport_through_box['AnthUpwelling 100 in']+dic_transport_through_box['AnthUpwelling 100 out'])

dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(
    dic_transport_through_box['surface_to_upperthermocline in']+dic_transport_through_box['surface_to_upperthermocline out'])
dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(
    dic_transport_through_box['anth_surface_to_upperthermocline in']+dic_transport_through_box['anth_surface_to_upperthermocline out'])


dic_transport_through_box['upperthermocline_sum']=dic_transport_through_box['upperthermocline_sum'].fillna(
    dic_transport_through_box['upper_to_lowerthermocline in']+dic_transport_through_box['upper_to_lowerthermocline out'])
dic_transport_through_box['upperthermocline_sum']=dic_transport_through_box['upperthermocline_sum'].fillna(
    dic_transport_through_box['anth_upper_to_lowerthermocline in']+dic_transport_through_box['anth_upper_to_lowerthermocline out'])

dic_transport_through_box['lowerthermocline_sum']=dic_transport_through_box['lowerthermocline_sum'].fillna(
    dic_transport_through_box['lowerthermocline_to_thermostad in']+dic_transport_through_box['lowerthermocline_to_thermostad out'])

dic_transport_through_box['lowerthermocline_sum']=dic_transport_through_box['lowerthermocline_sum'].fillna(
    dic_transport_through_box['anth_lowerthermocline_to_thermostad in']+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])

dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(dic_transport_through_box['CO2 flux in']+dic_transport_through_box['CO2 flux out'])
dic_transport_through_box['surface_sum']=dic_transport_through_box['surface_sum'].fillna(dic_transport_through_box['anth CO2 flux in']+dic_transport_through_box['anth CO2 flux out'])

#Total Sum 	Total Sum 100m 	Positive Sum 	Negative Sum
dic_transport_through_box['Total Sum']=dic_transport_through_box['Total Sum'].fillna(
    dic_transport_through_box['CO2 flux in']+dic_transport_through_box['CO2 flux out'])
dic_transport_through_box['Positive Sum']=dic_transport_through_box['Positive Sum'].fillna(
    dic_transport_through_box['CO2 flux in'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])
dic_transport_through_box['Negative Sum']=dic_transport_through_box['Negative Sum'].fillna(
    dic_transport_through_box['CO2 flux out'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])

dic_transport_through_box['Total Sum']=dic_transport_through_box['Total Sum'].fillna(
    dic_transport_through_box['anth CO2 flux in']+dic_transport_through_box['anth CO2 flux out'])
dic_transport_through_box['Positive Sum']=dic_transport_through_box['Positive Sum'].fillna(
    dic_transport_through_box['anth CO2 flux in'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])
dic_transport_through_box['Negative Sum']=dic_transport_through_box['Negative Sum'].fillna(
    dic_transport_through_box['anth CO2 flux out'])#+dic_transport_through_box['anth_lowerthermocline_to_thermostad out'])


dic_transport_through_box['Total Sum']=dic_transport_through_box['Total Sum'].fillna(
    dic_transport_through_box['anth_upper_to_lowerthermocline in']+dic_transport_through_box['anth_upper_to_lowerthermocline out'])
dic_transport_through_box['Total Sum']=dic_transport_through_box['Total Sum'].fillna(
    dic_transport_through_box['upper_to_lowerthermocline in']+dic_transport_through_box['upper_to_lowerthermocline out'])
    
    
    
# Need to do some calculations to fill the Positive Sum above thermocline and above thermostad ['Negative Sum']
# Also need to reverse some of the directions (ie E pacific?)
    
dic_transport_through_box=dic_transport_through_box.drop(['Upwelling 100 in','Upwelling 100 out','AnthUpwelling 100 in','AnthUpwelling 100 out'],axis=1)#,'Upwelling 23 out','Upwelling 24.5 in','Upwelling 24.5 out','Upwelling 26.3 in','Upwelling 26.3 out'],axis=1)

dic_transport_through_box=dic_transport_through_box.drop(['Upwelling 500 in','Upwelling 500 out','AnthUpwelling 500 in','AnthUpwelling 500 out'],axis=1)#,'Upwelling 23 out','Upwelling 24.5 in','Upwelling 24.5 out','Upwelling 26.3 in','Upwelling 26.3 out'],axis=1)
dic_transport_through_box=dic_transport_through_box.drop(['surface_to_upperthermocline in','surface_to_upperthermocline out','anth_surface_to_upperthermocline in','anth_surface_to_upperthermocline out','upper_to_lowerthermocline in','upper_to_lowerthermocline out','anth_upper_to_lowerthermocline in','anth_upper_to_lowerthermocline out'],axis=1)
dic_transport_through_box=dic_transport_through_box.drop(['lowerthermocline_to_thermostad in','lowerthermocline_to_thermostad out','anth_lowerthermocline_to_thermostad in','anth_lowerthermocline_to_thermostad out'],axis=1)
dic_transport_through_box=dic_transport_through_box.drop(['CO2 flux in','CO2 flux out','anth CO2 flux in','anth CO2 flux out'],axis=1)
dic_transport_through_box=dic_transport_through_box.dropna(thresh=1)

dic_transport_through_box.to_csv(f'dic_tx2{region}.csv')

t=pd.read_csv(f'dic_tx2{region}.csv').replace(0,np.nan)
merged=t.groupby('Unnamed: 0').mean().sort_values('Unnamed: 0').round(3)#agg(np.mean())

merged['enso_phase']=merged.index.str.split(' ').str[0]
merged.index=merged.index.str.split(' ').str[1:].str.join(' ')#.groupby('surface_sum')

merged.to_csv(f'dic_tx3{region}.csv')
merged



# Prepare Data for Plotting
dic_transport_through_box=pd.read_csv(f'dic_tx3{region}.csv',index_col=0).replace(0,np.nan)

names=dic_transport_through_box.index

new_tx=dic_transport_through_box[~names.str.lower().str.contains('old')]
new_tx_anth=new_tx[new_tx.index.str.lower().str.contains('anth')]
for anth in new_tx_anth.index:
    if anth in new_tx.index:
        new_tx=new_tx.drop(anth)
        
    
# Plotting
plt.figure(figsize=(16,16))
for i,col in enumerate(new_tx.columns[:-1]):
    ax=plt.subplot(4,2,i+1)
    df=new_tx[[col,'enso_phase']]#.plot.bar()
    df.pivot(columns='enso_phase').plot.bar(ax=ax)#plt.show()
    #df.drop('Upwelling 100').groupby('enso_phase').sum().T['EP'].values
    sum_of_rows=df.drop('Upwelling 100').groupby('enso_phase').sum()#
    #print(sum_of_rows
    plt.title(col)
    plt.xticks(rotation=45)
    plt.legend([f'CP {sum_of_rows.T["CP"].values[0].round(3)} PgAC',
            f'EP {sum_of_rows.T["EP"].values[0].round(3)} PgAC',
            f'Neutral {sum_of_rows.T["Neutral"].values[0].round(3)}PgC',
            f'Nina {sum_of_rows.T["Nina"].values[0].round(3)}PgC'])# Make sure this doesnt break anything
plt.tight_layout()
# Plotting Anth
plt.figure(figsize=(16,16))
for i, col in enumerate(new_tx_anth.columns[:-1]):
    ax=plt.subplot(4,2,i+1)
    df=new_tx_anth[[col,'enso_phase']]#.plot.bar()
    
    
    # CAN WE SOMEHOW MODIFY BAR COLOURS HERE?
    df.pivot(columns='enso_phase').plot.bar(ax=ax)#plt.show()
    #plt.legend(['CP','EP','Neutral','Nina'])# Make sure this doesnt break anything
    plt.title(col)
    sum_of_rows=df.drop('AnthUpwelling 100').groupby('enso_phase').sum()#
    #print(sum_of_rows)
    plt.xticks(rotation=45)
    plt.legend([f'CP {sum_of_rows.T["CP"].values[0].round(3)} PgAC',
                f'EP {sum_of_rows.T["EP"].values[0].round(3)} PgAC',
                f'Neutral {sum_of_rows.T["Neutral"].values[0].round(3)}PgC',
                f'Nina {sum_of_rows.T["Nina"].values[0].round(3)}PgC'])# Make sure this doesnt break anything
plt.tight_layout()
